refactor: split repository package by subdomains
